Add the ability to list the user's events and reminders for the specified number of days.
TutorPro currently does not have the ability to display to the user
their schedule for the next specified number of days.

Allowing the user to list their schedule for the next specified number
of days will make their schedule more efficient to read and allow
users to quickly find the information they need.

This commit will implement the function for users to list their
truncated schedule.